+++ empire/server/modules/bof/credentials/nanodump.py
@@ -10,6 +10,7 @@ class Module:
         params: dict,
         obfuscate: bool = False,
         obfuscation_command: str = "",
+        **kwargs,
     ):
         params_dict = {
             "Architecture": "x64",