+++ empire/server/modules/bof/situational_awareness/clipboard_window_inject_list.py
@@ -1,5 +1,4 @@
 import base64
-import json
 
 from empire.server.common.empire import MainMenu
 from empire.server.core.module_models import EmpireModule
@@ -14,15 +13,9 @@ class Module:
         params: dict,
         obfuscate: bool = False,
         obfuscation_command: str = "",
+        **kwargs,
     ):
-        bof_module = main_menu.modulesv2.modules["csharp_code_execution_runcoff"]
-        script_file = main_menu.modulesv2.dotnet_compiler.compile_task(
-            bof_module.compiler_yaml,
-            bof_module.name,
-            dot_net_version="net40",
-            confuse=obfuscate,
-        )
-
+        agent_language = kwargs.get("agent_language", "")
         script_path = main_menu.modulesv2.module_source_path / module.bof.x64
         bof_data = script_path.read_bytes()
         b64_bof_data = base64.b64encode(bof_data).decode("utf-8")
@@ -30,13 +23,9 @@ class Module:
         packer = Packer()
         packer.addint(0)
 
-        params_dict = {}
-        params_dict["Entrypoint"] = "go"
-        params_dict["File"] = b64_bof_data
-        params_dict["HexData"] = packer.getbuffer_data()
-
-        base64_json = base64.b64encode(json.dumps(params_dict).encode("utf-8")).decode(
-            "utf-8"
+        return main_menu.modulesv2.format_bof_output(
+            bof_data_b64=b64_bof_data,
+            hex_data=packer.getbuffer_data(),
+            agent_language=agent_language,
+            obfuscate=obfuscate,
         )
-
-        return f"{script_file}|,{base64_json}"